Add the homepage to SiteNavigation.__str__
A number of tests expected home to be included in the __str__ output for
SiteNavigation. Home isn't included as it isn't stored in
SiteNavigation.nav_items with the other pages.

This change adds it to the __str__ output and updates the expected
length for SiteNavigation.nav_items in the tests.