Updated setup.py to use entry_points['console_scripts'] instead of scripts.
This makes a "mkdocs" command on Windows/OSX/Linux without any platform-specific code in setup.py.

I think this is the preferred way to install a Python command with setuptools.

It is, at least, the method used by:

* [flake8](https://bitbucket.org/tarek/flake8/src/8ee94d1eeea36e835583b68c59682bfc3684fab9/setup.py?at=default)
* [coverage.py](https://bitbucket.org/ned/coveragepy/src/ca875e739048c3b333e862471b89d55e3c7b7170/setup.py?at=default)
* [Fabric](https://github.com/fabric/fabric/blob/master/setup.py)
* Many others

In the past, I had to change imports willy-nilly or add
"from __future__ import absolute_imports" to a bunch of files.
This is because I was renaming "mkdocs" to "mkdocs.py" instead of "main.py",
and the module-vs-script name clash was confusing imports from other files.