Page titles no longer contain the site name
If a page has a specific title, `page_title` will contain this and only
this. If the page has no title, or is the homepage, `page_title` will be
`None`. The `<title>` tag will always display at least the site name.

All of the themes' base.html files have been updated accordingly.