Preserve order of Python Markdown extensions
Python Markdown inclusion order can be important when including various
custom extensions.  Conflicts can arise as to where extensions get
inserted in the workflow.  By preserving order of how extensions are
defined in the settings file, this can allow a user to control
insertion order and work around these issues.  As Py26 supported is
noted to be removed in the future, I saw no reason to add additional
dependencies for OrderedDict.  Try catch statements can be removed when
Py26 support is removed.