Document the `--help` flag.
As of 40dc17e17e5, any mention of the `--help` flag had been removed from
the documentation. A breif summary has been added here. However, the output
of `--help` has not been included as that can change in the future and
updating the docs every time a option is updated become unnesecarily
duplicative (as the command line is self-documenting). This fixes #767.