Add tests for CLI options.
Previously we only tested that the CLI commands called the approiate function.
Now we also test that the correct options are passed. A few recently
introduced bugs were caught while building these tests (fixed in the
previous couple commits).

Left off tests for the `json` command as it is being deprecated (see #481).

Closes #1016 and #1014.