Break up libs and scripts into 2 seperate template blocks.
Some scripts need to execute after the page has finished loading.
To ensure that, the `scripts` block is placed at the bottom of the page.

However, if a user embeds scrpts within the body, and those scripts
rely in libs, then the libs need to already be loaded. By having
`libs` be a seperate block at the top of the page (int he header),
we can ensure that will work.

And we still want the libs in a block so a user can override them,
or even add to them (via `{{ super() }}`).

Fixes #956