Validate dev_addr config option.
Ensure reasonable error messages are returned on bad input.

This validates everything except the "host". We could use the ipaddress lib,
but we also should accept hostnames, not just IP addresses. However, by
validating the format and port, the address can more reliably be passed
to the server. The server does its own validation of the host and provides
its own reasonable error meessage.

As a side effect, we now support IPv6 as well as IPv4. Previously, the
port was not properly separated from an IPv6 address. Now that it is,
and the server already supports IPv6, we get support for free.

Fixes #1237.