Warn on relative paths in nav. (#1613)
While this feels like a feature change/addition, it is addressing a regression and is therefore being treated as a bug-fix.

Note that prior to 1.0 this was handled as part of the config validation. t is now handled after config validation completes. In strict mode, warnings issued in config validation interrupt the build after config validation is complete--causing the build to not complete. Warnings issued during the build (after config validation passes) allow the build to complete and only exit with a failed return code.

Closes #1604.

This also deletes relative_path_ext.py. That file was supposed to be removed some time ago. Not sure how it managed to return.