Skip duplicated error message when omitting the traceback
The new BuildError exception is good for skipping the traceback, but if it's skipped, now the two copies of the error mesage (one from logging, one from reaching SystemExit) immediately follow each other, which looks strange.
So, skip logging the exception's message when encountering a BuildError in _populate_page and _build_page.